+++ praw/__init__.py
@@ -1632,7 +1632,9 @@ class ModOnlyMixin(AuthenticatedReddit):
         return self.get_content(self.config['spam'] % six.text_type(subreddit),
                                 *args, **kwargs)
 
-    @decorators.restrict_access(scope=None, mod=True)
+    # OAuth scope modconfig for this method cannot be added until there has
+    # been an upstream change. See issue #174. Then the restriction will be
+    # @restrict_access('modconfig', mod=False, login=False)
     def get_stylesheet(self, subreddit):
         """Return the stylesheet and images for the given subreddit."""
         return self.request_json(self.config['stylesheet'] %